Mooving extension methods to a Syntax object
